@@ -13,29 +13,35 @@
     "build-storybook": "ng run firebase-app-hosting-angular:build-storybook"
   },
   "private": true,
+  "overrides": {
+    "undici": "6.19.2",
+    "undici-types": "6.19.2"
+  },
   "dependencies": {
-    "@angular/animations": "^17.3.1",
-    "@angular/cdk": "^17.3.10",
-    "@angular/common": "^17.3.1",
-    "@angular/compiler": "^17.3.1",
-    "@angular/core": "^17.3.1",
-    "@angular/fire": "^17.1.0",
-    "@angular/forms": "^17.3.1",
-    "@angular/material": "^17.3.10",
-    "@angular/platform-browser": "^17.3.1",
-    "@angular/platform-browser-dynamic": "^17.3.1",
-    "@angular/platform-server": "^17.3.1",
-    "@angular/router": "^17.3.1",
-    "@angular/ssr": "^17.3.3",
-    "@fortawesome/angular-fontawesome": "^0.14.1",
+    "@angular/animations": "^18.1.3",
+    "@angular/cdk": "^18.1.3",
+    "@angular/common": "^18.1.3",
+    "@angular/compiler": "^18.1.3",
+    "@angular/core": "^18.1.3",
+    "@angular/fire": "^18.0.1",
+    "@angular/forms": "^18.1.3",
+    "@angular/material": "^18.1.3",
+    "@angular/platform-browser": "^18.1.3",
+    "@angular/platform-browser-dynamic": "^18.1.3",
+    "@angular/platform-server": "^18.1.3",
+    "@angular/router": "^18.1.3",
+    "@angular/ssr": "^18.1.3",
+    "@fortawesome/angular-fontawesome": "^0.15.0",
     "@fortawesome/fontawesome-svg-core": "^6.4.2",
     "@fortawesome/free-brands-svg-icons": "^6.4.2",
     "@fortawesome/free-regular-svg-icons": "^6.4.2",
     "@fortawesome/free-solid-svg-icons": "^6.4.2",
+    "@ngrx/store": "^18.0.2",
+    "@ngrx/store-devtools": "^18.0.2",
     "express": "^4.18.2",
     "firebase": "^10.12.2",
     "firebase-functions": "^5.0.1",
-    "ngx-markdown": "^17.2.1",
+    "ngx-markdown": "^18.0.0",
     "primeflex": "^3.3.1",
     "primeicons": "^7.0.0",
     "primeng": "^17.18.2",
@@ -45,16 +51,17 @@
     "zone.js": "~0.14.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.3.1",
-    "@angular-eslint/builder": "17.5.2",
-    "@angular-eslint/eslint-plugin": "17.5.2",
-    "@angular-eslint/eslint-plugin-template": "17.5.2",
-    "@angular-eslint/schematics": "17.5.2",
-    "@angular-eslint/template-parser": "17.5.2",
-    "@angular/cli": "^17.3.1",
-    "@angular/compiler-cli": "^17.3.1",
+    "@angular-devkit/build-angular": "^18.1.3",
+    "@angular-eslint/builder": "18.2.0",
+    "@angular-eslint/eslint-plugin": "18.2.0",
+    "@angular-eslint/eslint-plugin-template": "18.2.0",
+    "@angular-eslint/schematics": "18.2.0",
+    "@angular-eslint/template-parser": "18.2.0",
+    "@angular/cli": "^18.1.3",
+    "@angular/compiler-cli": "^18.1.3",
     "@chromatic-com/storybook": "^1.6.0",
     "@compodoc/compodoc": "^1.1.25",
+    "@ngrx/eslint-plugin": "^18.0.2",
     "@storybook/addon-docs": "^8.1.11",
     "@storybook/addon-essentials": "^8.1.11",
     "@storybook/addon-interactions": "^8.1.11",
@@ -82,6 +89,6 @@
     "prettier": "^3.2.5",
     "prettier-eslint": "^16.3.0",
     "storybook": "^8.1.11",
-    "typescript": "~5.3.2"
+    "typescript": "~5.5.4"
   }
 }
